Revert "Make various code improvements"
This reverts commit 15571a71dc8cded8fc00301ff11371f38f1f3e0d.

+++ app/src/main/java/com/hellmund/primetime/ui/history/HistoryFragment.kt
@@ -64,18 +64,11 @@ class HistoryFragment : Fragment() {
     private fun render(viewState: HistoryViewState) {
         recyclerView.isVisible = viewState.isLoading.not()
         progressBar.isVisible = viewState.isLoading
-        adapter.submitList(viewState.data)
-
-        if (viewState.showRemoveFailedWarning) {
-            showToast(R.string.cant_remove_more_items)
-        }
+        adapter.update(viewState.data)
     }
 
     private fun onOpenDialog(movie: HistoryMovieViewEntity) {
-        val options = arrayOf(
-            getString(R.string.edit_rating),
-            getString(R.string.remove_from_history)
-        )
+        val options = getDialogOptions()
 
         requireContext().showItemsDialog(
             title = movie.title,
@@ -89,8 +82,19 @@ class HistoryFragment : Fragment() {
         )
     }
 
+    private fun getDialogOptions(): Array<String> {
+        return arrayOf(
+            getString(R.string.edit_rating),
+            getString(R.string.remove_from_history)
+        )
+    }
+
     private fun removeFromHistory(movie: HistoryMovieViewEntity) {
-        viewModel.dispatch(Action.Remove(movie))
+        if (adapter.canRemove()) {
+            viewModel.dispatch(Action.Remove(movie))
+        } else {
+            requireContext().showToast(R.string.cant_remove_more_items)
+        }
     }
 
     private fun openEditRatingDialog(movie: HistoryMovieViewEntity) {
